requirements.txt
tensorflow==2.16.1
keras==3.8.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.3
scikit-learn==1.3.2
keras-tuner==1.4.6
tensorflow-addons==0.21.0
transformers==4.40.0
binance-futures-connector==4.1.0
requests==2.31.0
tqdm==4.66.1
joblib==1.3.2
PyYAML==6.0.1
psutil==5.9.5
mypy==1.5.1
types-requests==2.31.0.2
optuna==3.1.0
sqlalchemy==2.0.25
alembic==1.13.1
plotly==5.18.0